feat: rename megcli to openapi-examples and update related configurations
