+++ .vscode/launch.json
@@ -5,11 +5,11 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "name": "Default debugger for megcli",
+            "name": "Default debugger for openapiexamplescli",
             "type": "go",
             "request": "launch",
             "mode": "auto",
-            "program": "./cmd/megcli/megcli.go",
+            "program": "./cmd/openapiexamplescli/openapiexamples.go",
             "cwd": "${workspaceFolder}",
             // please change these args before debugging. do not checkin changes.
             "args": [
